+++ changelog.md
@@ -1,3 +1,5 @@
+# 1.2.3
+- Fixed zero tank density when a `ModuleIgnitionTankController` is configured with only a single `PropellantConfig` with no specified `TankDensity`
 # 1.2.2
 - Added ignition potential computation using propellant configs when `AddedIgnitionPotential` is not specified on the `IGNITION_RESOURCE` node.
 - Fixed `FixedIgnitors` value not updating properly on `ModuleIgnitionEngineController` in `OnLoad`